Better Logging
Add `winston` as logging library and give it an exported configured log.
Add in logging in key places to ensure the site is starting properly.
Remove the random `console.log` line that was causing unhelpful output.
Capture the errors around fetching GitHub and Twitter info as likely
places to fail.