Cut down for rebuilding
Totally cut the app back to barebones to start building this back up
again. There's a lot of poorly written code here that needs to be
rewritten. I want to keep the git history around, so I'm not starting
totally over, but removing most everything so I can add the important
things back in.

Most importantly, the posts should stay, but they aren't going to be the
highlight of the blog. The homepage will show a new message and
direction.