Browser component (#55)
The entire point of making this blog be a custom website was to do
things like this. Build a component that mimics a brower, using an
`iframe` to load the external React project that the user is building.

This lets a user inspect the DOM right from their own browser! It also
allows them to select text or play around with the page. Hoping to do
more cool things like this moving forward.