Update MDX to v2
This was a big update. Had to change over the `next.config.js` file to
use `mjs` so it could read in the context's correctly. Then had to look
through all the files and make sure things didn't break too bad.

The good news is, I think this version fixes a lot of weird bugs with
MDX which should make writing easier.